+++ go.mod
@@ -3,7 +3,5 @@ module chatbot_golang
 go 1.18
 
 require (
-	github.com/gorilla/mux v1.8.0 // indirect
-	github.com/thedevsaddam/renderer v1.2.0 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
+
 )
